hehe

+++ ImpostorEngine2/source/Engine/IAudio.cpp
@@ -258,7 +258,7 @@ PUBLIC STATIC void IAudio::AudioCallback(void* data, uint8_t* stream, int len) {
                 bool deleted = false;
                 int bytes = audio->MusicStack[0]->Audio->RequestMoreData(AUDIO_SAMPLES, len);
                 if (bytes > 0) {
-                    SDL_MixAudioFormat(stream, audio->MusicStack[0]->Audio->Buffer, audio->DeviceFmt.format, (uint32_t)len, ((int)(audio->MusicStack[0]->Volume * (int)((audio->FadeOutTimer / audio->FadeOutTimerMax) * (audio->GlobalVolume * audio->MusicVolume))) >> 2) + 1);
+                    SDL_MixAudioFormat(stream, audio->MusicStack[0]->Audio->Buffer, audio->DeviceFmt.format, (uint32_t)len, ((int)(audio->MusicStack[0]->Volume * ((audio->FadeOutTimer / audio->FadeOutTimerMax) * (audio->GlobalVolume * audio->MusicVolume))) >> 2) + 1);
                 }
 				else if (bytes == -2) {
 					// waiting
